updated controllers

+++ src/main/java/com/github/ticketProject/javaSpringBootTemplate/service/UserService.java
@@ -2,6 +2,7 @@ package com.github.ticketProject.javaSpringBootTemplate.service;
 
 import com.github.ticketProject.javaSpringBootTemplate.model.Organization;
 import com.github.ticketProject.javaSpringBootTemplate.model.User;
+import com.github.ticketProject.javaSpringBootTemplate.model.UsersList;
 import com.github.ticketProject.javaSpringBootTemplate.repository.OrganizationRepository;
 import com.github.ticketProject.javaSpringBootTemplate.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,29 +40,28 @@ public class UserService {
         return userRepository.findAllByUsersListEquals(organization.getUsersList());
     }
 
-    public User createUser(Organization organization ,User user) {
+    public boolean createUser(UsersList usersList , User user) {
         // check if the username is already taken
         User userToCheck = userRepository.findByUsernameEquals(user.getUsername());
 
         // if we got back no User from the check save this new User
         if(userToCheck == null) {
             ZonedDateTime timeAsOfNow = ZonedDateTime.now();
-            user.setUsersList(organization.getUsersList());
             user.setDateCreated(timeAsOfNow);
             user.setLastModified(timeAsOfNow);
-            user.setFullName(user.getFirstName() + " " + user.getLastName());
+            usersList.addUser(user);
             userRepository.save(user);
-            return user;
+            return true;
         }
         // return the user saved for the HTTP response
-        return null;
+        return false;
     }
 
-    public boolean delete(Long idOfUser) {
+    public boolean delete(UsersList usersList, Long idOfUser) {
         // Use Spring's Crud Repository method to findById to get back an optional
         Optional<User> user = userRepository.findById(idOfUser);
         // Call the Crud Repository to deleteById if the Optional is Present
-        if(user.isPresent()) {
+        if(user.isPresent() && user.get().getUsersList().equals(usersList)) {
             userRepository.deleteById(idOfUser);
             return true;
         }
@@ -69,12 +69,12 @@ public class UserService {
         return false;
     }
 
-    public User editUser(Long idOfUser, User user) {
+    public User editUser(UsersList usersList, Long idOfUser, User user) {
         // first use findById to look for user
         Optional<User> findUser = this.findById(idOfUser);
 
         // Check the optional to see if anything is present then get the user object out else break out of this method
-        if(findUser.isPresent() ) {
+        if(findUser.isPresent() && findUser.get().getUsersList().equals(usersList)) {
             User returnedUser = findUser.get();
             // Set the changeable fields to the new fields if any change
             returnedUser.setFirstName(user.getFirstName());
@@ -93,21 +93,21 @@ public class UserService {
     }
 
 
-    public Organization addOrganizationToUser(Long idOfUser, Long idOfOrganization) {
-        // find the user first
-        Optional<User> user = findById(idOfUser);
-        Optional<Organization> organization = organizationRepository.findById(idOfOrganization);
-        // check if the user was found
-        if(user.isPresent() && organization.isPresent()) {
-            Organization organizationExist = organization.get();
-            User userExist = user.get();
-//            userExist.setOrganization(organizationExist);
-            userRepository.save(userExist);
-            // set the organization as the user
-            return organization.get();
-        }
-        return null;
-    }
+//    public Organization addOrganizationToUser(Long idOfUser, Long idOfOrganization) {
+//        // find the user first
+//        Optional<User> user = findById(idOfUser);
+//        Optional<Organization> organization = organizationRepository.findById(idOfOrganization);
+//        // check if the user was found
+//        if(user.isPresent() && organization.isPresent()) {
+//            Organization organizationExist = organization.get();
+//            User userExist = user.get();
+////            userExist.setOrganization(organizationExist);
+//            userRepository.save(userExist);
+//            // set the organization as the user
+//            return organization.get();
+//        }
+//        return null;
+//    }
 
     public boolean checkUsernameIsTaken(String username) {
         // check if the username is already taken
